Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/svg/animations/animate-path-animation-Qq-Tt-expected.txt


FAIL Test path animation where coordinate modes of start and end differ. You should see PASS messages assert_equals: expected "M -30 -30 Q 22.5 -30 22.5 0 T -30 30 Z" but got "M -30 -30 Q 30 -30 30 0 T -30 30 Z"
